Don't rely on the namespace context provided by the emitter
The StAX version built-in to JDK has a bug in its namespace context
implementation, where declared namespaces would bleed over to sibling,
for self-closing tags. This makes it unusable for DXML-25

We use our own bidirectional map impl for keeping track of the
namespace environment. This also simplifies the code somewhat, since
we don't have to work around interactions between the mutable context
and the emitter any more.