DXML-49: pervasively use pu-map
The ticket described a failure of using the builtin prefix xml:.
The reason for this commit being so large is, that in the process of
fixing it, I discovered that half of the system was still using an old
method of keeping track of the xmlns environment.

Properly fixing this simplified the whole library a lot and removed
lots of workarounds, obsoleted by pu-map.